pyproject.toml
[tool.poetry]
name = "loopfinder"
version = "0.1.0"
description = ""
authors = ["Mateus Junior <juniormateusknd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
jira = "^3.1.1"
pydantic = {version = "^1.9.0", allow-prereleases = true}
click = "^8.0.3"
rich = "^10.16.1"
injector = "^0.19.0"
questionary = "^1.10.0"
clipboard = "^0.0.4"
dependency-injector = "^4.37.0"
aiofiles = "^0.8.0"
psutil = "^5.9.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
mypy = "^0.950"
isort = "^5.10.1"
black = "^22.3.0"
tqdm = "^4.62.3"
types-PyYAML = "^6.0.1"
types-requests = "^2.27.7"
python-devtools = "^2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 89

[tool.pytest]
python_files = ['tests.py', 'test_*.py']

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.idea
    | _build
    | .pytest_cache
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
